// packages/cactus-plugin-htlc-eth-besu/src/main/typescript/generated/openapi/typescript-axios/api.ts
export interface NewContractObj {
    /**
     * Contract address
     * @type {string}
     * @memberof NewContractObj
     */
    contractAddress: string;
    /**
     * 
     * @type {number}
     * @memberof NewContractObj
     */
    inputAmount?: number;
    /**
     * 
     * @type {number}
     * @memberof NewContractObj
     */
    outputAmount: number;
    /**
     * 
     * @type {number}
     * @memberof NewContractObj
     */
    expiration: number;
    /**
     * 
     * @type {string}
     * @memberof NewContractObj
     */
    hashLock: string;
    /**
     * 
     * @type {string}
     * @memberof NewContractObj
     */
    receiver?: string;
    /**
     * 
     * @type {string}
     * @memberof NewContractObj
     */
    outputNetwork: string;
    /**
     * 
     * @type {string}
     * @memberof NewContractObj
     */
    outputAddress: string;
    /**
     * Web3SigningCredentialType
     * @type {string}
     * @memberof NewContractObj
     */
    credentialType: string;
    /**
     * KeychainId for the keychain plugin
     * @type {string}
     * @memberof NewContractObj
     */
    keychainId: string;
    /**
     * connectorId for the connector besu plugin
     * @type {string}
     * @memberof NewContractObj
     */
    connectorId: string;
    /**
     * 
     * @type {string}
     * @memberof NewContractObj
     */
    accountRef: string;
    /**
     * 
     * @type {string}
     * @memberof NewContractObj
     */
    privateKeyRef: string;
    /**
     * 
     * @type {string | number}
     * @memberof NewContractObj
     */
    gas?: string | number;
}
/**
 * 
 * @export
 * @interface RefundReq
 */
export interface RefundReq {
    /**
     * Contract address
     * @type {string}
     * @memberof RefundReq
     */
    contractAddress: string;
    /**
     * Web3SigningCredentialType
     * @type {string}
     * @memberof RefundReq
     */
    credentialType: string;
    /**
     * KeychainId for the keychain plugin
     * @type {string}
     * @memberof RefundReq
     */
    keychainId: string;
    /**
     * connectorId for the connector besu plugin
     * @type {string}
     * @memberof RefundReq
     */
    connectorId: string;
    /**
     * 
     * @type {string}
     * @memberof RefundReq
     */
    accountRef: string;
    /**
     * 
     * @type {string}
     * @memberof RefundReq
     */
    privateKeyRef: string;
    /**
     * 
     * @type {string | number}
     * @memberof RefundReq
     */
    gas?: string | number;
}
/**
 * 
 * @export
 * @enum {string}
 */
export enum Web3SigningCredentialType {
    CACTUSKEYCHAINREF = 'CACTUS_KEYCHAIN_REF',
    GETHKEYCHAINPASSWORD = 'GETH_KEYCHAIN_PASSWORD',
    PRIVATEKEYHEX = 'PRIVATE_KEY_HEX',
    NONE = 'NONE'
}

/**
 * 
 * @export
 * @interface WithdrawReq
 */
export interface WithdrawReq {
    /**
     * 
     * @type {string}
     * @memberof WithdrawReq
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof WithdrawReq
     */
    secret: string;
    /**
     * Contract address
     * @type {string}
     * @memberof WithdrawReq
     */
    contractAddress: string;
    /**
     * Web3SigningCredentialType
     * @type {string}
     * @memberof WithdrawReq
     */
    credentialType: string;
    /**
     * KeychainId for the keychain plugin
     * @type {string}
     * @memberof WithdrawReq
     */
    keychainId: string;
    /**
     * connectorId for the connector besu plugin
     * @type {string}
     * @memberof WithdrawReq
     */
    connectorId: string;
    /**
     * 
     * @type {string}
     * @memberof WithdrawReq
     */
    accountRef: string;
    /**
     * 
     * @type {string}
     * @memberof WithdrawReq
     */
    privateKeyRef: string;
    /**
     * 
     * @type {string | number}
     * @memberof WithdrawReq
     */
    gas?: string | number;
}